Move PhpFilePathsPlugin to separate file
